+++ src/formalities/__init__.py
@@ -5,5 +5,5 @@ formalities
 a portable formal logic framework for enabling validated reasoning capabilities in agentic LLM's
 """
 
-__version__ = "0.0.2"
+__version__ = "0.0.3"
 __author__ = "Joel Yisrael"